CLN: Consolidate groupby.value_counts tests
